+++ ts/features/pn/screens/__test__/PaidPaymentScreen.test.tsx
@@ -1,39 +1,31 @@
-import { cleanup } from "@testing-library/react-native";
-import configureMockStore from "redux-mock-store";
-import { PaidPaymentScreen } from "../PaidPaymentScreen";
-import PN_ROUTES from "../../navigation/routes";
-import { GlobalState } from "../../../../store/reducers/types";
-import { applicationChangeState } from "../../../../store/actions/application";
+import { createStore } from "redux";
 import { appReducer } from "../../../../store/reducers";
+import { applicationChangeState } from "../../../../store/actions/application";
+import { preferencesDesignSystemSetEnabled } from "../../../../store/actions/persistedPreferences";
 import { renderScreenWithNavigationStoreContext } from "../../../../utils/testWrapper";
+import PN_ROUTES from "../../navigation/routes";
+import { PaidPaymentScreen } from "../PaidPaymentScreen";
+import { GlobalState } from "../../../../store/reducers/types";
 
 describe("PaidPaymentScreen", () => {
-  // Needed to avoid `ReferenceError: You are trying to `import` a file after the Jest environment has been torn down.`
-  afterEach(cleanup);
-  it("should render with back button, title, help button, paid banner, notice code and creditor tax id", () => {
-    const component = generateComponent(
-      generateNoticeCode(),
-      generateCreditorTaxId()
-    );
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-  it("should render with back button, title, help button, paid banner and notice code but no creditor tax id", () => {
-    const component = generateComponent(generateNoticeCode());
-    expect(component.toJSON()).toMatchSnapshot();
+  it("should match snapshot", () => {
+    const screen = renderScreen();
+    expect(screen.toJSON()).toMatchSnapshot();
   });
 });
 
-const generateNoticeCode = () => "018011988086479497";
-const generateCreditorTaxId = () => "00000000009";
-
-const generateComponent = (noticeCode: string, creditorTaxId?: string) => {
-  const globalState = appReducer(undefined, applicationChangeState("active"));
-  const store = configureMockStore<GlobalState>()(globalState);
+const renderScreen = () => {
+  const initialState = appReducer(undefined, applicationChangeState("active"));
+  const designSystemState = appReducer(
+    initialState,
+    preferencesDesignSystemSetEnabled({ isDesignSystemEnabled: true })
+  );
+  const store = createStore(appReducer, designSystemState as any);
 
   return renderScreenWithNavigationStoreContext<GlobalState>(
     PaidPaymentScreen,
     PN_ROUTES.CANCELLED_MESSAGE_PAID_PAYMENT,
-    { noticeCode, creditorTaxId },
+    { noticeCode: "012345678912345670", creditorTaxId: "01234567890" },
     store
   );
 };
